+++ internal/s3api/s3_multipart_handlers.go
@@ -33,7 +33,7 @@ func (s *S3Gateway) InitiateMultipartUpload(w http.ResponseWriter, r *http.Reque
 	bucket := mux.Vars(r)["bucket"]
 	key := mux.Vars(r)["key"]
 
-	err := s.client.InitMultipartUpload(bucket, key, uploadID)
+	err := s.multiPartStore.InitMultipartUpload(bucket, key, uploadID)
 	if err != nil {
 		model.WriteErrorResponse(w, r, model.ErrInternalError)
 		return
@@ -96,7 +96,7 @@ func (s *S3Gateway) UploadPart(w http.ResponseWriter, r *http.Request) {
 		Closer: r.Body,
 	}
 
-	etag, err := s.client.UploadPart(bucket, key, uploadID, partNum, limitedBody)
+	etag, err := s.multiPartStore.UploadPart(bucket, key, uploadID, partNum, limitedBody)
 	if err != nil {
 		if errors.Is(err, client.ErrUploadNotFound) || errors.Is(err, client.ErrUploadCompleted) {
 			model.WriteErrorResponse(w, r, model.ErrNoSuchUpload)
@@ -143,7 +143,7 @@ func (s *S3Gateway) CompleteMultipartUpload(w http.ResponseWriter, r *http.Reque
 	}
 
 	sortedPartNumbers := parsePartNumbers(parts)
-	etag, err := s.client.CompleteMultipartUpload(bucket, key, uploadID, sortedPartNumbers)
+	etag, err := s.multiPartStore.CompleteMultipartUpload(bucket, key, uploadID, sortedPartNumbers)
 	if err != nil {
 		model.WriteErrorResponse(w, r, model.ErrInternalError)
 		return
@@ -168,7 +168,7 @@ func (s *S3Gateway) AbortMultipartUpload(w http.ResponseWriter, r *http.Request)
 		model.WriteErrorResponse(w, r, model.ErrNoSuchUpload)
 		return
 	}
-	err := s.client.AbortMultipartUpload(bucket, key, uploadID)
+	err := s.multiPartStore.AbortMultipartUpload(bucket, key, uploadID)
 	if err != nil {
 		model.WriteErrorResponse(w, r, model.ErrInternalError)
 		return
@@ -205,7 +205,7 @@ func (s *S3Gateway) ListParts(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	meta, err := s.client.ListParts(bucket, key, uploadID)
+	meta, err := s.multiPartStore.ListParts(bucket, key, uploadID)
 	if err != nil {
 		if errors.Is(err, client.ErrUploadNotFound) || errors.Is(err, client.ErrUploadCompleted) {
 			model.WriteErrorResponse(w, r, model.ErrNoSuchUpload)
